backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
asgiref==3.11.0
attrs==25.4.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.1.8
Django==5.2.8
django-cors-headers==4.9.0
djangorestframework==3.16.1
edge-tts==7.2.3
frozenlist==1.8.0
gTTS==2.5.4
idna==3.11
multidict==6.7.0
pillow==12.0.0
propcache==0.4.1
pydub==0.25.1
requests==2.32.5
sqlparse==0.5.4
tabulate==0.9.0
typing_extensions==4.15.0
urllib3==2.5.0
yarl==1.22.0
gunicorn==23.0.0
whitenoise==6.8.2
